move fake impl to its own file
